Refactor training code into modular structure
- Split 592-line train.py into focused modules:
  * config.py: Configuration constants
  * preprocessing.py: Data preprocessing utilities
  * models.py: Model training functions
  * visualization.py: Plotting functions
  * train.py: Main orchestration (now 303 lines)

- Removed emoji usage throughout
- Improved code organization and maintainability
- All functionality preserved and tested
- Old train.py backed up as train_old.py
